chore(client/collection): call add and update api in bookmark detail store
