+++ client/src/app/core/auth/auth.provider.ts
@@ -3,13 +3,13 @@ import { EnvironmentProviders, Provider, inject } from "@angular/core";
 import { provideHttpClient, withInterceptors } from "@angular/common/http";
 import { provideAppInitWithConfigAsync } from "@core/config";
 import { AuthService } from "./data-access/auth.service";
-import { authInterceptor } from "./interceptors/auth.interceptor";
+import { fakeBackendInterceptor } from "./interceptors/fake-backend.interceptor";
 
 export const provideAuth = (): EnvironmentProviders | Provider => [
   provideHttpClient(
     withInterceptors([
-      authInterceptor,
-      // fakeBackendInterceptor
+      // authInterceptor,
+      fakeBackendInterceptor,
     ])
   ),
   provideAppInitWithConfigAsync(() => {
