+++ client/src/app/features/bookmark/bookmark.routes.ts
@@ -1,15 +1,21 @@
 import { Routes } from "@angular/router";
 import { BookmarkManagementComponent } from "./containers/bookmark-management/bookmark-management.component";
-import { BookmarkStore, provideBookmarkApi, provideCrawlApi } from "./data-access";
+import {
+  BookmarkDetailStore,
+  BookmarkStore,
+  provideBookmarkMockApi,
+  provideCrawlApi,
+} from "./data-access";
 
 export const bookmarkRoutes: Routes = [
   {
     path: "",
     providers: [
-      provideBookmarkApi(),
-      // provideBookmarkMockApi(),
+      // provideBookmarkApi(),
+      provideBookmarkMockApi(),
       provideCrawlApi(),
       BookmarkStore,
+      BookmarkDetailStore,
     ],
     component: BookmarkManagementComponent,
   },
