+++ client/src/app/features/bookmark/containers/bookmark-management/bookmark-management.component.ts
@@ -4,13 +4,19 @@ import {
   Component,
   Injector,
   OnInit,
+  ViewContainerRef,
   inject,
 } from "@angular/core";
 import { ActivatedRoute } from "@angular/router";
+import { BookmarkDetailDialogComponent } from "@bookmark/components/bookmark-detail-dialog/bookmark-detail-dialog.component";
+import { BookmarkDetailFacade } from "@bookmark/components/bookmark-detail-dialog/bookmark-detail.facade";
 import { BookmarkListComponent } from "@bookmark/components/bookmark-list/bookmark-list.component";
+import { PadDialogService } from "@shared/ui";
+import { isNotFalsy, isNotNil } from "@shared/utils";
 import { HlmButtonDirective } from "@spartan-ng/ui-button-helm";
 import { HlmIconComponent, provideIcons } from "@spartan-ng/ui-icon-helm";
 import { HlmH4Directive } from "@spartan-ng/ui-typography-helm";
+import { filter, map } from "rxjs";
 import { BookmarkFacade } from "./bookmark-management.facade";
 
 // Generics
@@ -37,6 +43,7 @@ const lucideCirclePlus = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 2
     BookmarkListComponent,
   ],
   providers: [
+    BookmarkDetailFacade,
     BookmarkFacade,
     provideIcons({
       lucideCirclePlus,
@@ -45,8 +52,11 @@ const lucideCirclePlus = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 2
 })
 export class BookmarkManagementComponent implements OnInit {
   private readonly _injector = inject(Injector);
+  private readonly _vcr = inject(ViewContainerRef);
+  private readonly _dialogService = inject(PadDialogService);
   private readonly _facade = inject(BookmarkFacade);
   private readonly _route = inject(ActivatedRoute);
+
   $collection = this._facade.$collection;
   $entities = this._facade.$entities;
   $pagination = this._facade.$pagination;
@@ -66,10 +76,27 @@ export class BookmarkManagementComponent implements OnInit {
   }
 
   onAdd(): void {
-    this._facade.add();
+    const data$ = this._dialogService
+      .open(BookmarkDetailDialogComponent, {
+        closeOnBackdropClick: false,
+        contentClass: "max-w-[30rem]",
+        vcr: this._vcr,
+      })
+      .closed$.pipe(filter(isNotNil));
+
+    this._facade.add(data$);
   }
 
   onDelete(id: string): void {
-    this._facade.delete(id);
+    const id$ = this._dialogService
+      .openConfirmDialog({
+        description: `This action cannot be undone. It will permanently delete your bookmark, from our servers`,
+        confirmText: `Delete bookmark`,
+      })
+      .pipe(
+        filter(isNotFalsy),
+        map(() => id)
+      );
+    this._facade.delete(id$);
   }
 }
