+++ client/src/app/features/bookmark/data-access/bookmark-management.store.ts
@@ -1,0 +1,110 @@
+import { inject } from "@angular/core";
+import { ServerSideError } from "@core/http";
+import { patchState, signalStore, withMethods } from "@ngrx/signals";
+import {
+  addEntity,
+  removeEntity,
+  setAllEntities,
+  withEntities,
+} from "@ngrx/signals/entities";
+import { rxMethod } from "@ngrx/signals/rxjs-interop";
+import {
+  setError,
+  setFulfilled,
+  setPending,
+  withPagination,
+  withStatus,
+} from "@shared/data-access";
+import { BookmarkMessage } from "@shared/enums";
+import { BookmarkVM } from "@shared/models";
+import { ToastService } from "@shared/services";
+import { PadDialogService } from "@shared/ui";
+import { isNotNil, prefix } from "@shared/utils";
+import { EMPTY, catchError, filter, map, pipe, switchMap, tap } from "rxjs";
+import { injectBookmarkApi } from ".";
+
+export const BookmarkStore = signalStore(
+  withStatus(),
+  withEntities<BookmarkVM>(),
+  withPagination(),
+  withMethods(store => {
+    const bookmarkApi = injectBookmarkApi();
+    const dialogService = inject(PadDialogService);
+    const toastService = inject(ToastService);
+
+    return {
+      ...store,
+      findAll: rxMethod<string>(
+        pipe(
+          tap(() => {
+            patchState(store, setPending());
+          }),
+          switchMap(collectionId =>
+            bookmarkApi
+              .findAll({
+                collectionId: collectionId,
+                currentPage: store.currentPage(),
+                pageSize: store.pageSize(),
+                keyword: "",
+              })
+              .pipe(
+                tap({
+                  next: res => {
+                    patchState(store, setAllEntities(res.result.items), setFulfilled());
+                  },
+                  error: err => {
+                    // TODO: using logger service
+                    patchState(store, setError(err));
+                  },
+                }),
+                catchError(() => EMPTY)
+              )
+          )
+        )
+      ),
+      create: rxMethod<BookmarkVM>(
+        pipe(
+          prefix(() => patchState(store, setPending())),
+          tap({
+            next: data => patchState(store, addEntity(data)),
+            error: err => toastService.error(err.message),
+          }),
+          catchError(() => EMPTY)
+        )
+      ),
+      delete: rxMethod<string>(
+        pipe(
+          map(id => store.entities().find(item => item.id === id)),
+          filter(isNotNil),
+          switchMap(data =>
+            bookmarkApi.delete(data.id).pipe(
+              prefix(() => patchState(store, setPending())),
+              tap({
+                next: () => {
+                  patchState(store, removeEntity(data.id), setFulfilled());
+                  toastService.success(`Bookmark “${data.title}“ was deleted`);
+                },
+                error: err => {
+                  let message = `Bookmark “${data.title}” could not be deleted`;
+                  if (err instanceof ServerSideError) {
+                    switch (err.message) {
+                      case BookmarkMessage.NotExist:
+                        message = `Bookmark “${data.title}” to be deleted does not exist`;
+                        break;
+
+                      default:
+                        break;
+                    }
+                  }
+                  patchState(store, setError(err));
+                  toastService.error(message);
+                },
+              }),
+              catchError(() => EMPTY)
+            )
+          )
+        )
+      ),
+    };
+  })
+);
