+++ client/src/app/features/bookmark/data-access/index.ts
@@ -1,3 +1,4 @@
+export * from "./bookmark-detail.store";
+export * from "./bookmark-management.store";
 export * from "./bookmark.provider";
-export * from "./bookmark.store";
 export * from "./crawl.provider";
