+++ client/src/app/features/shell/containers/layout/layout.component.html
@@ -15,7 +15,7 @@
 <hlm-toaster />
 @if ($loading()) {
   <div
-    class="w-screen h-screen fixed z-[99] top-0 left-0 bg-white bg-opacity-60 flex justify-center items-center">
+    class="w-screen h-screen fixed z-[9999] top-0 left-0 bg-white bg-opacity-60 flex justify-center items-center">
     <hlm-spinner />
   </div>
 }
